+++ modules/sd_models.py
@@ -974,16 +974,38 @@ def load_diffuser(checkpoint_info=None, op='model', revision=None): # pylint: di
             move_model(sd_model, devices.device)
         timer.load.record("move")
 
-        if shared.opts.ipex_optimize:
-            sd_model = sd_models_compile.ipex_optimize(sd_model)
-
-        if ('Model' in shared.opts.cuda_compile and shared.opts.cuda_compile_backend != 'none'):
-            sd_model = sd_models_compile.compile_diffusers(sd_model)
-        timer.load.record("compile")
-
     except Exception as e:
         log.error(f"Load {op}: {e}")
         errors.display(e, "Model")
+
+    try:
+        if shared.opts.ipex_optimize:
+            sd_model = sd_models_compile.ipex_optimize(sd_model)
+
+        if (shared.opts.cuda_compile_backend != 'none') and len(shared.opts.cuda_compile) > 0:
+            if 'components' in shared.opts.cuda_compile_options:
+                sd_model = sd_models_compile.compile_diffusers(sd_model, apply_to_components=True)
+            else:
+                if 'Model' in shared.opts.cuda_compile:
+                    if hasattr(sd_model, "unet"):
+                        sd_model.unet = sd_models_compile.compile_diffusers(sd_model.unet, apply_to_components=False)
+                    if hasattr(sd_model, "transformer"):
+                        sd_model.transformer = sd_models_compile.compile_diffusers(sd_model.transformer, apply_to_components=False)
+                if 'TE' in shared.opts.cuda_compile:
+                    if hasattr(sd_model, "text_encoder"):
+                        sd_model.text_encoder = sd_models_compile.compile_diffusers(sd_model.text_encoder, apply_to_components=False)
+                    if hasattr(sd_model, "text_encoder_2"):
+                        sd_model.text_encoder_2 = sd_models_compile.compile_diffusers(sd_model.text_encoder_2, apply_to_components=False)
+                    if hasattr(sd_model, "text_encoder_3"):
+                        sd_model.text_encoder_3 = sd_models_compile.compile_diffusers(sd_model.text_encoder_3, apply_to_components=False)
+                if 'VAE' in shared.opts.cuda_compile:
+                    if hasattr(sd_model, "vae"):
+                        sd_model.vae = sd_models_compile.compile_diffusers(sd_model.vae, apply_to_components=False)
+
+        timer.load.record("compile")
+    except Exception as e:
+        log.error(f"Compile {op}: {e}")
+        errors.display(e, "Compile")
 
     if shared.opts.diffusers_offload_mode != 'balanced':
         devices.torch_gc(force=True, reason='load')