+++ modules/sd_offload.py
@@ -234,7 +234,6 @@ class OffloadHook(accelerate.hooks.ModelHook):
             return False
         return True
 
-    @torch.compiler.disable
     def pre_forward(self, module, *args, **kwargs):
         _id = id(module)
 
@@ -298,7 +297,6 @@ class OffloadHook(accelerate.hooks.ModelHook):
         self.last_pre = _id
         return args, kwargs
 
-    @torch.compiler.disable
     def post_forward(self, module, output):
         if self.last_post != id(module):
             self.last_post = id(module)